requirements.txt
flask
flask_socketio
pandas
pymongo
scikit-learn
requests
python-dotenv
